+++ roles/jd-server/src/lib/mod.rs
@@ -9,7 +9,10 @@ use roles_logic_sv2::{
     errors::Error, parsers::PoolMessages as JdsMessages, utils::CoinbaseOutput as CoinbaseOutput_,
 };
 use serde::Deserialize;
-use std::convert::{TryFrom, TryInto};
+use std::{
+    convert::{TryFrom, TryInto},
+    time::Duration,
+};
 use stratum_common::bitcoin::{Script, TxOut};
 
 pub type Message = JdsMessages<'static>;
@@ -63,4 +66,35 @@ pub struct Configuration {
     pub core_rpc_port: u16,
     pub core_rpc_user: String,
     pub core_rpc_pass: String,
+    #[serde(deserialize_with = "duration_from_toml")]
+    pub mempool_update_timeout: Duration,
+}
+
+fn duration_from_toml<'de, D>(deserializer: D) -> Result<Duration, D::Error>
+where
+    D: serde::Deserializer<'de>,
+{
+    #[derive(Deserialize)]
+    struct Helper {
+        unit: String,
+        value: u64,
+    }
+
+    let helper = Helper::deserialize(deserializer)?;
+    match helper.unit.as_str() {
+        "seconds" => Ok(Duration::from_secs(helper.value)),
+        "secs" => Ok(Duration::from_secs(helper.value)),
+        "s" => Ok(Duration::from_secs(helper.value)),
+        "milliseconds" => Ok(Duration::from_millis(helper.value)),
+        "millis" => Ok(Duration::from_millis(helper.value)),
+        "ms" => Ok(Duration::from_millis(helper.value)),
+        "microseconds" => Ok(Duration::from_micros(helper.value)),
+        "micros" => Ok(Duration::from_micros(helper.value)),
+        "us" => Ok(Duration::from_micros(helper.value)),
+        "nanoseconds" => Ok(Duration::from_nanos(helper.value)),
+        "nanos" => Ok(Duration::from_nanos(helper.value)),
+        "ns" => Ok(Duration::from_nanos(helper.value)),
+        // ... add other units as needed
+        _ => Err(serde::de::Error::custom("Unsupported duration unit")),
+    }
 }
