+++ roles/jd-server/src/main.rs
@@ -1,4 +1,4 @@
-#![allow(special_module_name)]
+/* #![allow(special_module_name)]
 use crate::lib::{mempool, status, Configuration};
 use async_channel::unbounded;
 use codec_sv2::{StandardEitherFrame, StandardSv2Frame};
@@ -13,7 +13,7 @@ use error_handling::handle_result;
 use stratum_common::bitcoin::{Script, TxOut};
 use tracing::{error, info, warn};
 
-use crate::lib::mempool;
+//use crate::lib::mempool;
 use roles_logic_sv2::utils::Mutex;
 use std::{sync::Arc, time::Duration};
 use tokio::{select, task};
@@ -39,7 +39,18 @@ pub struct Configuration {
     pub core_rpc_pass: String,
     #[serde(deserialize_with = "duration_from_toml")]
     pub mempool_update_timeout: Duration,
-}
+} */
+#![allow(special_module_name)]
+use crate::lib::{mempool, status, Configuration};
+use async_channel::unbounded;
+use error_handling::handle_result;
+use roles_logic_sv2::utils::Mutex;
+use std::sync::Arc;
+use tokio::{select, task};
+use tracing::{error, info, warn};
+mod lib;
+
+use lib::job_declarator::JobDeclarator;
 
 mod args {
     use std::path::PathBuf;
@@ -101,35 +112,6 @@ mod args {
             };
             Ok(Self { config_path })
         }
-    }
-}
-
-fn duration_from_toml<'de, D>(deserializer: D) -> Result<Duration, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    #[derive(Deserialize)]
-    struct Helper {
-        unit: String,
-        value: u64,
-    }
-
-    let helper = Helper::deserialize(deserializer)?;
-    match helper.unit.as_str() {
-        "seconds" => Ok(Duration::from_secs(helper.value)),
-        "secs" => Ok(Duration::from_secs(helper.value)),
-        "s" => Ok(Duration::from_secs(helper.value)),
-        "milliseconds" => Ok(Duration::from_millis(helper.value)),
-        "millis" => Ok(Duration::from_millis(helper.value)),
-        "ms" => Ok(Duration::from_millis(helper.value)),
-        "microseconds" => Ok(Duration::from_micros(helper.value)),
-        "micros" => Ok(Duration::from_micros(helper.value)),
-        "us" => Ok(Duration::from_micros(helper.value)),
-        "nanoseconds" => Ok(Duration::from_nanos(helper.value)),
-        "nanos" => Ok(Duration::from_nanos(helper.value)),
-        "ns" => Ok(Duration::from_nanos(helper.value)),
-        // ... add other units as needed
-        _ => Err(serde::de::Error::custom("Unsupported duration unit")),
     }
 }
